+++ src/main/java/ru/hogwarts/school/repositories/StudentRepository.java
@@ -19,7 +19,7 @@ public interface StudentRepository extends JpaRepository<Student, Long> {
     int countStudent();
 
     @Query(value = "SELECT AVG(age) FROM student", nativeQuery = true)
-    int countStudentsAverageAge();
+    double countStudentsAverageAge();
 
     @Query(value = "SELECT * FROM student ORDER BY id DESC LIMIT 5", nativeQuery = true)
     List<Student> getLastFiveStudents();
